backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alpaca-py==0.30.1
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==25.12.0
boto3==1.42.5
botocore==1.42.5
build==1.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.3.7
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
curl_cffi==0.13.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
fastuuid==0.14.0
feedparser==6.0.11
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.55.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jq==1.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
librt==0.7.3
langgraph==0.2.45
langchain==0.3.15
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-google-genai==2.0.8
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.3.1
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy==1.19.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.99.9
orjson==3.11.5
overrides==7.7.0
packaging>=23.2,<25
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.37.2
stripe==14.0.1
sympy==1.14.0
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.25.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
